+++ lib/utopia/project/linkify.rb
@@ -3,7 +3,7 @@
 # Released under the MIT License.
 # Copyright, 2020-2024, by Samuel Williams.
 
-require 'decode/syntax/rewriter'
+require "decode/syntax/rewriter"
 
 module Utopia
 	module Project
@@ -24,14 +24,14 @@ module Utopia
 			
 			def link_to(definition, text)
 				XRB::Builder.fragment do |builder|
-					builder.inline('a', href: @base.link_for(definition), title: definition.qualified_name) do
+					builder.inline("a", href: @base.link_for(definition), title: definition.qualified_name) do
 						builder.text(text)
 					end
 				end
 			end
 			
 			def apply(output = XRB::Builder.new)
-				output.inline('code', class: "language-#{@language.name}") do
+				output.inline("code", class: "language-#{@language.name}") do
 					super
 				end