+++ pages/source/controller.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
 # Released under the MIT License.
-# Copyright, 2020-2022, by Samuel Williams.
+# Copyright, 2020-2024, by Samuel Williams.
 
 prepend Actions
 
-on '**/*/index' do |request, path|
+on "**/*/index" do |request, path|
 	@lexical_path = path.components.dup
 	# Remove the last "index" part:
 	@lexical_path.pop