+++ template/preload.rb
@@ -3,4 +3,4 @@
 # Released under the MIT License.
 # Copyright, 2020-2023, by Samuel Williams.
 
-require 'utopia/project'
+require "utopia/project"