+++ test/utopia/project.rb
@@ -3,7 +3,7 @@
 # Released under the MIT License.
 # Copyright, 2020-2023, by Samuel Williams.
 
-require 'utopia/project'
+require "utopia/project"
 
 describe Utopia::Project do
 	it "has a version number" do